pyproject.toml
[tool.poetry]
name = "ctk_api"
version = "0.1.0"
description = "A beautiful description."
authors = ["Reinder Vos de Wael <reinder.vosdewael@childmind.org>"]
license = "LGPL-2.1"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
pydantic = "^2.3.0"
python-docx = "^0.8.11"
pydantic-settings = "^2.0.3"
python-multipart = "^0.0.6"
uvicorn = "^0.23.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
mypy = "^1.4.0"
black = "^23.3.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
pytest-cov = "^4.1.0"
httpx = "^0.25.0"

[tool.poetry.group.docs.dependencies]
pdoc = "^14.0.0"

[tool.poetry.extras]
docs = ["pdoc"]

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.mypy]
ignore_missing_imports = true
